+++ src/ReconNess.Core/Providers/ILogsProvider.cs
@@ -1,0 +1,34 @@
+﻿using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace ReconNess.Core.Providers
+{
+    /// <summary>
+    /// The interface ILogsProvider
+    /// </summary>
+    public interface ILogsProvider
+    {
+        /// <summary>
+        /// Obtain the list of logs file
+        /// </summary>
+        /// <param name="cancellationToken">Notification that operations should be canceled</param>
+        /// <returns>The list of logs file</returns>
+        IEnumerable<string> GetLogfiles(CancellationToken cancellationToken);
+
+        /// <summary>
+        /// Obtain the log file content
+        /// </summary>
+        /// <param name="logFileSelected">The log file</param>
+        /// <param name="cancellationToken">Notification that operations should be canceled</param>
+        /// <returns>The log file content</returns>
+        Task<string> ReadLogfile(string logFileSelected, CancellationToken cancellationToken);
+
+        /// <summary>
+        /// Clean the log file
+        /// </summary>
+        /// <param name="logFileSelected">The log file</param>
+        /// <param name="cancellationToken">Notification that operations should be canceled</param>
+        void CleanLogfile(string logFileSelected, CancellationToken cancellationToken);
+    }
+}
